pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "openai-agents>=0.2.11",
    "pinecone>=7.3.0",
    "python-dotenv>=1.1.1",
    "supabase>=2.18.1",
    "uvicorn[standard]>=0.35.0",
]
